+++ core/src/com/tann/dice/util/TextWisp.java
@@ -40,7 +40,7 @@ public class TextWisp extends BasicLay{
 //        batch.setColor(Colours.withAlpha(Colours.dark, alp));
 //        Draw.fillRectangle(batch,getX()-100,getY()-100,200,200);
 		batch.setColor(c.r, c.g, c.b, alp);
-		TannFont.font.drawString(batch, text, getX(), getY());
+		TannFont.font.drawString(batch, text, (int)getX(), (int) getY());
 	}
 	
 }
